Moved Product class to it's own file, changed imports as needed.
